Create Mode class, update docstrings and imputers
