+++ creme/stats/__init__.py
@@ -7,6 +7,7 @@ from .kurtosis import Kurtosis
 from .max import Max
 from .mean import Mean
 from .min import Min
+from .mode import Mode
 from .n_unique import NUnique
 from .ptp import PeakToPeak
 from .skew import Skew
@@ -20,6 +21,7 @@ __all__ = [
     'Max',
     'Mean',
     'Min',
+    'Mode',
     'NUnique',
     'PeakToPeak',
     'Skew',
